+++ src/RhinoInside.Revit.GH/Components/Views/Identity.cs
@@ -5,9 +5,10 @@ using ARDB = Autodesk.Revit.DB;
 namespace RhinoInside.Revit.GH.Components.Views
 {
   using External.DB.Extensions;
+  using Grasshopper.Kernel.Parameters;
 
-  [ComponentVersion(introduced: "1.0", updated: "1.2.1")]
-  public class ViewIdentity : Component
+  [ComponentVersion(introduced: "1.0", updated: "1.7")]
+  public class ViewIdentity : ZuiComponent
   {
     public override Guid ComponentGuid => new Guid("B0440885-4AF3-4890-8E84-3BC2A8342B9F");
     public override GH_Exposure Exposure => GH_Exposure.primary;
@@ -23,21 +24,42 @@ namespace RhinoInside.Revit.GH.Components.Views
     )
     { }
 
-    protected override void RegisterInputParams(GH_InputParamManager manager)
+    protected override ParamDefinition[] Inputs => inputs;
+    static readonly ParamDefinition[] inputs =
     {
-      manager.AddParameter(new Parameters.View(), "View", "View", string.Empty, GH_ParamAccess.item);
-    }
+      ParamDefinition.Create<Parameters.View>("View", "V", "View to query")
+    };
 
-    protected override void RegisterOutputParams(GH_OutputParamManager manager)
+    protected override ParamDefinition[] Outputs => outputs;
+    static readonly ParamDefinition[] outputs =
     {
-      manager.AddParameter(new Parameters.Param_Enum<Types.ViewDiscipline>(), "Discipline", "D", "View discipline", GH_ParamAccess.item);
-      manager.AddParameter(new Parameters.Param_Enum<Types.ViewFamily>(), "Family", "F", "View family", GH_ParamAccess.item);
-      manager.AddTextParameter("Name", "N", "View name", GH_ParamAccess.item);
-      manager.AddParameter(new Parameters.View(), "Template", "T", "View template", GH_ParamAccess.list);
-      manager.AddBooleanParameter("Is Template", "IT", "View is template", GH_ParamAccess.item);
-      manager.AddBooleanParameter("Is Printable", "IP", "View is printable", GH_ParamAccess.item);
-      manager.AddBooleanParameter("Is Assembly", "IA", "View is assembly", GH_ParamAccess.item);
-      manager.AddParameter(new Parameters.AssemblyInstance(), "Assembly", "A", "Assembly instance that owns the assembly view", GH_ParamAccess.item);
+      ParamDefinition.Create<Parameters.Param_Enum<Types.ViewDiscipline>>("Discipline", "D", "View discipline", relevance: ParamRelevance.Primary),
+      ParamDefinition.Create<Parameters.Param_Enum<Types.ViewFamily>>("View Family", "VF", "View family"),
+      ParamDefinition.Create<Param_String>("View Name", "VN", "View name"),
+      ParamDefinition.Create<Param_Boolean>("Is Dependent", "D", "View is dependent", relevance: ParamRelevance.Primary),
+      ParamDefinition.Create<Parameters.View>("View Dependency", "VD", "View depedency", relevance: ParamRelevance.Occasional),
+      ParamDefinition.Create<Param_Boolean>("Is Template", "IT", "View is template", relevance: ParamRelevance.Primary),
+      ParamDefinition.Create<Parameters.View>("View Template", "T", "View template", relevance: ParamRelevance.Occasional),
+      ParamDefinition.Create<Param_Boolean>("Is Assembly", "IA", "View is assembly", relevance: ParamRelevance.Primary),
+      ParamDefinition.Create<Parameters.AssemblyInstance>("Associated Assembly", "AA", "The assembly the view is associated with", relevance: ParamRelevance.Occasional),
+      ParamDefinition.Create<Param_Boolean>("Is Printable", "IP", "View is printable", relevance: ParamRelevance.Primary),
+    };
+
+    public override void AddedToDocument(GH_Document document)
+    {
+      if (Params.Output<IGH_Param>("Family") is IGH_Param family)
+        family.Name = "View Family";
+
+      if (Params.Output<IGH_Param>("Name") is IGH_Param name)
+        name.Name = "View Name";
+
+      if (Params.Output<IGH_Param>("Template") is IGH_Param template)
+        template.Name = "View Template";
+
+      if (Params.Output<IGH_Param>("Assembly") is IGH_Param assembly)
+        assembly.Name = "Associated Assembly";
+
+      base.AddedToDocument(document);
     }
 
     protected override void TrySolveInstance(IGH_DataAccess DA)
@@ -51,13 +73,15 @@ namespace RhinoInside.Revit.GH.Components.Views
       else
         DA.SetData("Discipline", null);
 
-      DA.SetData("Family", view.GetViewFamily());
-      DA.SetData("Name", view.Name);
-      DA.SetData("Template", new Types.View(view.Document, view.ViewTemplateId));
-      DA.SetData("Is Template", view.IsTemplate);
-      DA.SetData("Is Printable", view.CanBePrinted);
-      DA.SetData("Is Assembly", view.IsAssemblyView);
-      DA.SetData("Assembly", new Types.AssemblyInstance(view.Document, view.AssociatedAssemblyInstanceId));
+      Params.TrySetData(DA, "View Family", () => view.GetViewFamily());
+      Params.TrySetData(DA, "View Name", () => view.Name);
+      Params.TrySetData(DA, "Is Dependent", () => view.GetPrimaryViewId() != ARDB.ElementId.InvalidElementId);
+      Params.TrySetData(DA, "View Depedency", () => Types.View.FromElementId(view.Document, view.GetPrimaryViewId()));
+      Params.TrySetData(DA, "Is Template", () => view.IsTemplate);
+      Params.TrySetData(DA, "View Template", () => Types.View.FromElementId(view.Document, view.ViewTemplateId));
+      Params.TrySetData(DA, "Is Assembly", () => view.IsAssemblyView);
+      Params.TrySetData(DA, "Associated Assembly", () => new Types.AssemblyInstance(view.Document, view.AssociatedAssemblyInstanceId));
+      Params.TrySetData(DA, "Is Printable", () => view.CanBePrinted);
     }
   }
 }